+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+- Fix mandatory fields on a Tab block not being enforced when updating an item.
 - Fix administrators losing access to a block's configuration after setting a profile to "no access" on that block.
 - Fix dependency conflict with GLPI core by no longer vendoring symfony/deprecation-contracts and symfony/polyfill-ctype.
 - Fix default field values not being applied when fields are empty on creation